requirements.txt
Flask==3.0.1
numpy==1.25.1
python_Levenshtein==0.23.0
scipy==1.12.0
spacy==3.6.1
torch==2.0.1
transformers==4.30.2
